+++ app/scripts/fhir/sdc-export-common.js
@@ -825,7 +825,7 @@ function addCommonSDCExportFns(ns) {
             answer = { "valueString" : itemValue };
           }
           else if (!jQuery.isEmptyObject(itemValue)) {
-            var answerCoding = this._setIfHasValue(null, 'system', LForms.Util.getCodeSystem(itemValue.codeSystem));
+            var answerCoding = this._setIfHasValue(null, 'system', LForms.Util.getCodeSystem(itemValue.system));
             answerCoding = this._setIfHasValue(answerCoding, 'code', itemValue.code);
             answerCoding = this._setIfHasValue(answerCoding, 'display', itemValue.text);
             answer = this._setIfHasValue(null, 'valueCoding', answerCoding);