+++ app/scripts/lib/lforms-hl7.js
@@ -378,8 +378,8 @@ LForms.HL7 = (function() {
         rtn = this.delimiters.component.repeat(8) + itemVal.text;
       }
       else {
-        var answerCS = !itemVal.codeSystem ? "" : (itemVal.codeSystem === 'LOINC' || itemVal.codeSystem === LOINC_URI) ?
-            this.LOINC_CS : itemVal.codeSystem;
+        var answerCS = !itemVal.system ? "" : (itemVal.system === 'LOINC' || itemVal.system === LOINC_URI) ?
+            this.LOINC_CS : itemVal.system;
         rtn = code + this.delimiters.component +
           itemVal.text + this.delimiters.component + answerCS;
       }