+++ test/protractor/spec/lforms_skiplogic.spec.js
@@ -25,6 +25,7 @@ describe('skip logic', function() {
     ff.cneTriggerSrc1.sendKeys(protractor.Key.TAB);
 
     expect(ff.dobIfLivingYes.isDisplayed()).toBe(true);
+    expect(ff.dobIfLivingYesB.isPresent()).toBeFalsy(); // trigger value has no 'system' while answers have 'system'
     expect(ff.ageIfLivingAnswered.isDisplayed()).toBe(true);
     expect(ff.deathCauseIfLivingNo.isPresent()).toBeFalsy();
     expect(ff.ageDeathIfLivingNotAnswered.isPresent()).toBeFalsy();
@@ -38,6 +39,7 @@ describe('skip logic', function() {
     ff.cneTriggerSrc1.sendKeys(protractor.Key.TAB);
 
     expect(ff.dobIfLivingYes.isPresent()).toBeFalsy();
+    expect(ff.dobIfLivingYesB.isPresent()).toBeFalsy(); // trigger value has no 'system' while answers have 'system'
     expect(ff.ageIfLivingAnswered.isDisplayed()).toBe(true);
     expect(ff.deathCauseIfLivingNo.isDisplayed()).toBe(true);
     expect(ff.ageDeathIfLivingNotAnswered.isPresent()).toBeFalsy();
@@ -48,9 +50,31 @@ describe('skip logic', function() {
     ff.cneTriggerSrc1.clear();
 
     expect(ff.dobIfLivingYes.isPresent()).toBeFalsy();
+    expect(ff.dobIfLivingYesB.isPresent()).toBeFalsy();
     expect(ff.ageIfLivingAnswered.isPresent()).toBeFalsy();
     expect(ff.deathCauseIfLivingNo.isPresent()).toBeFalsy();
     expect(ff.ageDeathIfLivingNotAnswered.isDisplayed()).toBe(true);
+  });
+
+  it('should show/hide elements when "system" value is set correctly in trigger if answered Yes', function() {
+    // clear the answer to Living
+    ff.cneTriggerSrc2.click();
+    ff.cneTriggerSrc2.sendKeys(protractor.Key.ARROW_DOWN);
+    ff.cneTriggerSrc2.sendKeys(protractor.Key.TAB);
+
+    expect(ff.dobIfLivingYes2C.isPresent()).toBeFalsy(); // trigger value has 'system' while answers have no 'system'
+    expect(ff.dobIfLivingYes2D.isDisplayed()).toBe(true);
+    expect(ff.deathCauseIfLivingNoB.isPresent()).toBeFalsy();
+
+    ff.cneTriggerSrc2.click();
+    ff.cneTriggerSrc2.sendKeys(protractor.Key.ARROW_DOWN);
+    ff.cneTriggerSrc2.sendKeys(protractor.Key.ARROW_DOWN);
+    ff.cneTriggerSrc2.sendKeys(protractor.Key.TAB);
+
+    expect(ff.dobIfLivingYes2C.isPresent()).toBeFalsy(); // trigger value has 'system' while answers have no 'system'
+    expect(ff.dobIfLivingYes2D.isPresent()).toBeFalsy();
+    expect(ff.deathCauseIfLivingNoB.isDisplayed()).toBe(true);
+
   });
 
   it('should show a sibling and two items in a sibling section', function() {