+++ test/protractor/spec/lforms_testpage.po.js
@@ -191,9 +191,16 @@ var TestPage = function() {
 
       cneTriggerSrc1: element(by.id('/54139-1-cnesrc-1/1')),
       dobIfLivingYes: element(by.id('/54139-1-cnesrc-1/54124-3/1/1')),
+      dobIfLivingYesB: element(by.id('/54139-1-cnesrc-1/54124-3b/1/1')),
       ageIfLivingAnswered: element(by.id('/54139-1-cnesrc-1/54141-7/1/1')),
       deathCauseIfLivingNo: element(by.id('/54139-1-cnesrc-1/54112-8/1/1')),
-      ageDeathIfLivingNotAnswered: element(by.id('/54139-1-cnesrc-1/54113-6/1/1'))
+      ageDeathIfLivingNotAnswered: element(by.id('/54139-1-cnesrc-1/54113-6/1/1')),
+
+      cneTriggerSrc2: element(by.id('/54139-1-cnesrc-2/1')),
+      dobIfLivingYes2C: element(by.id('/54139-1-cnesrc-2/54124-3c/1/1')),
+      dobIfLivingYes2D: element(by.id('/54139-1-cnesrc-2/54124-3d/1/1')),
+      deathCauseIfLivingNoB: element(by.id('/54139-1-cnesrc-2/54112-8b/1/1'))
+
     },
 
     HL7GeneticPanel: {