+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [21.2.1] 2020-03-05
+### Changed
+- Changed answer's 'codeSystem' to be 'system'. Legacy form data with 'codeSystem'
+  are converted into 'system' when a form a loaded.
+
 ## [21.2.0] 2020-03-03
 ### Added
 - LForms form definitions now contain an lformsVersion attribute indicating the